move HttpErrorException to internal.exception package
